beat_testing: using align
Align removes or adds extra notes.  Removing a note does not yet incur
a penalty.  Added notes are given a large error value, to cause an
overall failure.

This function isn't used very well do to the fact that input is read on
a counter (for notes).  This means extra notes will cause the last note
not to be read and missing a note will cause the input routine to wait
for the last key press.

Considering a sampling based approached.