beat_testing:  create an abstraction over basic fuctions
A 'pattern' is just a list of ones and/or zeros.  Each one represents
a note and the zero represents silence.

The basic data structure is now a 'pattern'.  This moves the conversion
functions into the right places and gives a clean high level interface.
Instead of wrapping the basic functions with other conversion functions,
they now take a 'pattern' and a duration.

We still need to convert read-samples to return a pattern and a duration
and analyze should accept a pattern and duration as the second argument
(it now takes the output of read-samples, which is a list of deltas.