beat_testing: clean up the play function
The new wait-for-rest-notes functions is called as the base case in
the main timming loop.  Before it was clumsily tacked on to the end of
the play function. Also, the play function only accounted for up to one
rest note at the end of the pattern.  Now it will properly wait for any
amount of final rests.